requirements.txt
anyio==3.6.2
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.6
Django==4.2
django-admin-interface==0.24.2
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-js-asset==2.0.0
django-render-partial==0.4
exceptiongroup==1.1.1
filelock==3.11.0
flet==0.5.2
flet-core==0.5.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
instaloader==4.9.6
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
platformdirs==3.2.0
pycparser==2.21
PySocks==1.7.1
repath==0.9.0
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
watchdog==2.3.1
websocket-client==1.5.1
websockets==10.4
wsproto==1.2.0
